# internal/serve/openapi.yaml
openapi: 3.1.0
info:
  title: imas CLI API
  description: |
    Local HTTP API served by `imas serve`. Proxies requests to the imas
    farmer over NATS and serves the embedded web UI.

    This API is intended for local use only (default: localhost:7505).
    All endpoints under `/api/v1/` are JSON-over-HTTP proxies to the
    NATS-based farmer API.

    ## Response Envelope

    Most endpoints that proxy through NATS return responses wrapped in
    a standard envelope:

    ```json
    {"result": <data>}
    ```

    On error:

    ```json
    {"error": "message describing the problem"}
    ```

    The schemas below describe the `result` payload for each endpoint.
  version: 0.2.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: imas
    url: https://github.com/yogzblr/imas
servers:
  - url: http://localhost:7505
    description: Default local imas serve address
paths:
...
  /api/v1/recipes:
    get:
      operationId: listRecipes
      summary: List all recipes
      description: |
        Returns a list of all recipes available on the farmer, including
        their dot-notation names, file paths, and sizes.
      tags:
        - recipes
      responses:
        "200":
          description: Recipe list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NATSEnvelope"
              examples:
                success:
                  value:
                    result:
                      recipes:
                        - name: base.webserver
                          path: base/webserver.imas
                          size: 1234
                        - name: monitoring.prometheus
                          path: monitoring/prometheus.imas
                          size: 2048
        "502":
          $ref: "#/components/responses/NATSError"
  /api/v1/recipes/{id...}:
    get:
      operationId: getRecipe
      summary: Get a recipe by path
      description: |
        Returns the contents of a recipe file by its dot-notation name.
        Uses a wildcard path parameter to support nested paths
        (e.g., `base.webserver` or `base/webserver`).
      tags:
        - recipes
      parameters:
        - name: id
          in: path
          required: true
          description: Recipe name in dot-notation (e.g., `base.webserver`) or path form
          schema:
            type: string
      responses:
        "200":
          description: Recipe contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NATSEnvelope"
              examples:
                success:
                  value:
                    result:
                      name: base.webserver
                      path: base/webserver.imas
                      content: |
                        nginx:
                          pkg.installed:
                            - name: nginx
                          service.running:
                            - name: nginx
                            - enable: true
                      size: 1234
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/NATSError"
...
components:
...
  schemas:
...
    NATSEnvelope:
      type: object
      description: |
        Standard NATS proxy response envelope. On success, `result`
        contains the response data. On error, `error` contains a
        message string.
      properties:
        result:
          description: Response data (structure varies by endpoint — see endpoint examples)
        error:
          type: string
          description: Error message (present only on failure)
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: missing id parameter
  responses:
    BadRequest:
      description: Bad request — missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NATSError:
      description: NATS communication error — farmer unreachable or returned an error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...